fix(db): cast death_day to integer in onDeath batch UPDATE
* fix(tests): correct BUG-12 test assertion (alive_count vs _aliveIds)

processDisaster() removes dead individuals from _aliveIds but does NOT
decrement worldState.alive_count directly — that is set at the start of
each tick from alive.length. Updated the BUG-12 test group to assert on
_aliveIds.size and _newDeadThisTick.length instead of worldState.alive_count.

All 328 server tests now pass.


* fix(db): cast death_day to integer in onDeath batch UPDATE

PostgreSQL infers untyped VALUES parameters as 'text' when the column
type cannot be determined from context. This caused every death persist
to fail with:
  column "death_day" is of type integer but expression is of type text

Fix: add ::integer cast on v.dd in the SET clause so Postgres knows
to coerce the value before assignment.


---------

+++ server/src/api/simulationManager.js
@@ -113,7 +113,7 @@ class SimulationManager {
       const values = dead.flatMap(d => [d.id, d.death_day ?? null, d.death_cause ?? null]);
       const rows = dead.map((_, i) => `($${i * 3 + 1}, $${i * 3 + 2}, $${i * 3 + 3})`).join(', ');
       query(
-        `UPDATE individuals SET alive = false, death_day = v.dd, death_cause = v.dc
+        `UPDATE individuals SET alive = false, death_day = v.dd::integer, death_cause = v.dc
          FROM (VALUES ${rows}) AS v(id, dd, dc)
          WHERE individuals.id = v.id::uuid`,
         values
